Replace ContainerSpec with ContainerTests.
